Create user _galera in galera-arbitrator-4 instead of galera-4
The galera-4 binary package is a plugin for mariadb: this library
already runs in the context of the mariadbd daemon: typically as mysql
user and doesn't require any specific user for itself.

The standalone garbd arbitrator daemon usually runs on a system where
mariadb-server and thus galera-4 is not installed. Because of this,
systemd fails to start garbd because it doesn't find the _galera user
now added in garbd.service's configuration.

Simply move the postinst script creating _galera from galera-4 to
galera-arbitrator-4 to fix this.

Signed-off-by: Adel Belhouane <bugs.a.b@free.fr>