Salsa CI: Remove mariadb-10.1 and mariadb-10.3 upgrade tests
Both the Stretch and Buster upgrades are now failing with probably
unsolvable dependency chains related to new systemd and apt versions.

As the Debian releases are quite old, and MariaDB 10.3 has been out of
standard support for nearly a year, address the failing CI by simply
removing the tests as they are probably no longer needed.

    $ test -f /etc/init.d/mariadb || apt-get install -q --yes mariadb-server
    Reading package lists...
    Building dependency tree...
    Reading state information...
    Some packages could not be installed. This may mean that you have
    requested an impossible situation or if you are using the unstable
    distribution that some required packages have not yet been created
    or been moved out of Incoming.
    The following information may help to resolve the situation:
    The following packages have unmet dependencies:
     libsystemd0 : Depends: libc6 (>= 2.39) but 2.38-11 is to be installed
    E: Unable to correct problems, you have held broken packages.